Move media embedding to a new Embed extension
- Allow any extension as long as a MIME type of audio/video is provided
- Remove fallback message. Only show a download link with alt text if
  provided an if not, the link will show the URL.
