+++ mkdocs.yml
@@ -153,6 +153,7 @@ markdown_extensions:
   - pymdownx.tabbed:
       alternate_style: true
   - pymdownx.saneheaders:
+  - pymdownx.embed:
 
 extra:
   social:
